Added StraightDistance function to Drivetrain. Added DriveStraightReverse function to Drivetrain. Changed DriveStraight to (Driver X).
Much like ETA function, in the StraightDistance function you set a set
point and a set angle, not a time, distance, and angle. May be a step
down, or it may be the function we call if we want to just get to a
distance.

DriveStraightReverse is the same as DriveStraightForward but in reverse.

New Buttons:
	Driver:
		(X) DriveStraightForward
		(B) DriveStraightReverse
		(LB) ResetFlags and ResetPIDs
		(Y) StraightDistance